refactor: cli dirs and filenames
