+++ packages/cli/src/cli-error.ts
