+++ packages/cli/src/commands/generate.ts
@@ -25,15 +25,15 @@ import { Command } from "commander";
 import { relative } from "pathe";
 import color from "picocolors";
 import packageJson from "../../package.json" with { type: "json" };
+import { CLIError } from "../cli-error.js";
 import { ERROR_MESSAGES } from "../constants/error-messages.js";
-import { getProjectInfo } from "../detection/framework.js";
-import { prepareTokensForCLI } from "../pipelines/prepare-tokens-for-cli.js";
+import { handleError } from "../handle-error.js";
+import { prepareTokens } from "../prepare-tokens.js";
+import { getProjectInfo } from "../project/framework.js";
 import { warningBoxWithBadge } from "../prompts/box-with-badge.js";
 import { intro, label, outro } from "../prompts/common.js";
 import { log } from "../prompts/log.js";
-import { getMarkupFiles, readMarkupSources } from "../scanning/markup.js";
-import { CLIError } from "../types/errors.js";
-import { handleError } from "../utils/handle-error.js";
+import { getMarkupFiles, readMarkupSources } from "../scan-markup.js";
 import { startWatcher } from "../watch/watcher.js";
 
 const CSS_BANNER = `/* Generated by @sugarcube-sh/cli v${packageJson.version} */\n\n`;
@@ -305,7 +305,7 @@ async function runGeneration(
     options: GenerateAllCSSOptions = {}
 ): Promise<GenerationResult> {
     clearMatchCache();
-    const { trees, resolved, warnings } = await prepareTokensForCLI(config);
+    const { trees, resolved, warnings } = await prepareTokens(config);
     const output = await generateAllCSS(trees, resolved, config, options);
     return { output, warnings };
 }
