+++ packages/cli/src/installation/components.ts
@@ -1,10 +1,10 @@
 import { mkdir, writeFile } from "node:fs/promises";
 import { basename, join } from "pathe";
+import { CLIError } from "../cli-error.js";
 import { getRegistryFiles } from "../registry/client.js";
-import { CLIError } from "../types/errors.js";
+import { resolveTree } from "../registry/dependency-tree.js";
 import type { InstallComponentsOptions, InstallComponentsResult } from "../types/install.js";
 import type { RegistryItem } from "../types/registry.js";
-import { resolveTree } from "../utils/resolve-dependencies.js";
 import { installDependencies } from "./dependencies.js";
 
 interface RegistryFileWithContent {
