+++ packages/cli/src/installation/dependencies.ts
@@ -1,6 +1,6 @@
 import { execa } from "execa";
+import { CLIError } from "../cli-error.js";
 import { ERROR_MESSAGES } from "../constants/error-messages.js";
-import { CLIError } from "../types/errors.js";
 import type { PackageManager } from "../types/install.js";
 
 export async function installDependencies(
