+++ packages/cli/src/prepare-tokens.ts
@@ -1,11 +1,11 @@
 import { loadTokens, resolveTokens } from "@sugarcube-sh/core";
 import type { InternalConfig, TokenPipelineSource } from "@sugarcube-sh/core";
-import { ERROR_MESSAGES } from "../constants/error-messages.js";
-import { CLIError } from "../types/errors.js";
+import { CLIError } from "./cli-error.js";
+import { ERROR_MESSAGES } from "./constants/error-messages.js";
 
 // CLI needs special error handling so we wrap the token processing pipeline
 // I considered separating the two but this felt like the right abstraction
-export async function prepareTokensForCLI(
+export async function prepareTokens(
     validatedConfig: InternalConfig,
     memoryData?: Record<string, { set?: string; content: string }>
 ) {
