+++ packages/cli/src/project/is-package-installed.ts
