+++ packages/cli/src/project/package-manager.ts
