--- packages/cli/src/utils/directory-resolver.ts
@@ -1,8 +1,0 @@
-import { resolve } from "pathe";
-
-export function resolveDirectoryFromFlag(flagValue: string): {
-    absolute: string;
-} {
-    const absolute = resolve(process.cwd(), flagValue);
-    return { absolute };
-}
