--- packages/cli/src/validation/preflight-init.ts
@@ -1,10 +1,0 @@
-import { configFileExists } from "@sugarcube-sh/core";
-import { ERROR_MESSAGES } from "../constants/error-messages.js";
-import { errorBoxWithBadge } from "../prompts/box-with-badge.js";
-
-export async function preflightInit(): Promise<void> {
-    if (configFileExists()) {
-        errorBoxWithBadge(ERROR_MESSAGES.CONFIG_EXISTS(), {});
-        process.exit(1);
-    }
-}
