+++ packages/cli/src/write-token-files.ts
